added logic for checking duplicates

+++ bot.py
@@ -15,24 +15,30 @@ auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
 auth.set_access_token(acess_token, acess_secret)
 api = tweepy.API(auth)
 
-mention=api.mentions_timeline(1)[0]
-screen_name = mention.author.screen_name
-id = mention.id_str
-text=mention.text
+mentions=api.mentions_timeline(count=20)
 
-filename = "temp.jpg"
-url = "https://preview.redd.it/d6i4k905snv61.jpg?width=640&height=424&crop=smart&auto=webp&s=bac4a75feb6d00e50139c0b54b1a96aaff877ccf"
-request = requests.get(url, stream=True)
+check=[]
 
-if request.status_code == 200:
-    with open(filename, 'wb') as image:
-        for chunk in request:
-            image.write(chunk)
-    image=api.media_upload(filename)
-    os.remove(filename)
+for mention in mentions:
+    if mention.id not in check:
+        check.append(mention.id)
+        screen_name = mention.author.screen_name
+        id = mention.id_str
+        text=mention.text
 
-media_id=image.media_id_string
-api.update_status(status='@'+screen_name,in_reply_to_status_id=id,media_ids=[media_id])
+        filename = "temp.jpg"
+        url = "https://preview.redd.it/d6i4k905snv61.jpg?width=640&height=424&crop=smart&auto=webp&s=bac4a75feb6d00e50139c0b54b1a96aaff877ccf"
+        request = requests.get(url, stream=True)
+
+        if request.status_code == 200:
+            with open(filename, 'wb') as image:
+                for chunk in request:
+                    image.write(chunk)
+            image=api.media_upload(filename)
+            os.remove(filename)
+
+        media_id=image.media_id_string
+        api.update_status(status='@'+screen_name,in_reply_to_status_id=id,media_ids=[media_id])
